pyproject.toml
[project]
name = "Lilex"
authors = [
    {name = "Mikhael Khrustik", email = "misha@myrt.co"},
]
description = "The font for developers."
readme = "README.md"
requires-python = ">=3.13"
version = "v0.0.0"
dependencies = [
    "arrrgs>=3.1.0",
    "glyphslib==6.11.6",
    "pyyaml==6.0.3",
    "fontmake==3.11.0",
    "fonttools==4.60.1",
    "cu2qu==1.6.7.post2",
    "gftools==0.9.91",
    "ruff==0.13.2",
    "pylint==3.3.8",
    "colored==2.3.1",
    "pylance==0.37.0",
    "fontbakery==1.1.0",
    "dehinter==4.0.0",
    "shaperglot==1.2.0",
    "youseedee==0.7.0",
    "pyarrow==21.0.0",
    "rich>=14.0.0",
]

[tool.ruff]
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
lint.ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
